+++ app/api/views/product_view.py
@@ -3,7 +3,7 @@ import http
 from flask import Blueprint, request
 
 from app.api.views.cart_view import scope_decorator
-from app.exceptions.exceptions import AccessDeniedException, ProductNotFoundException
+from app.exceptions.exceptions import AccessDeniedException, ProductNotFoundException, ProductExistsException
 from app.orm.models.product.product_model import ProductModel
 from app.orm.repository import product_repository
 from app.orm.schemas.query.product.product_query import ProductQueryParam
@@ -17,9 +17,8 @@ product_blueprint = Blueprint('product_blueprint', __name__, url_prefix="/produc
 def find_all():
     query_param: ProductQueryParam = ProductQueryParam.as_obj(request.args.to_dict(flat=True))
     products: list[ProductModel] = product_repository.filtered_sorted_product_list(query_param)
-    paginator = query_param.paginator.copy(update={"total": product_repository.count()})
     return ProductResponseSchema(items=[ProductSchema.from_orm(product).dict() for product in products],
-                                 paginator=paginator).dict(), http.HTTPStatus.OK
+                                 paginator=query_param.paginator).dict(), http.HTTPStatus.OK
 
 
 @product_blueprint.route('/<int:id>')
@@ -35,6 +34,9 @@ def single_product(id: int):
 def new_product(scope):
     if "create product" not in scope:
         raise AccessDeniedException
+    product = product_repository.find_by_name(request.json['name'])
+    if product:
+        raise ProductExistsException
     product = product_repository.create(ProductRequestSchema.parse_obj(request.json))
     return ProductSchema.from_orm(product).json(), http.HTTPStatus.CREATED
 
@@ -55,5 +57,8 @@ def update_product(scope, id: int):
 def delete_product(scope, id: int):
     if "delete product" not in scope:
         raise AccessDeniedException
+    product = product_repository.find(id)
+    if not product:
+        raise ProductNotFoundException
     product_repository.delete(id)
     return {}, http.HTTPStatus.NO_CONTENT
