+++ app/api/views/user_view.py
@@ -3,7 +3,7 @@ import http
 from flask import Blueprint, jsonify, request
 
 from app.api.views.cart_view import scope_decorator
-from app.exceptions.exceptions import AccessDeniedException
+from app.exceptions.exceptions import AccessDeniedException, UserNotFoundException, UserExistsException
 from app.orm.models.user.user_model import UserModel
 from app.orm.repository import user_repository, cart_repository
 from app.orm.schemas.request.cart.cart import CartSchema
@@ -28,6 +28,8 @@ def single_user(scope, id: int):
     if "get user" not in scope:
         raise AccessDeniedException
     user = user_repository.find(id)
+    if not user:
+        raise UserNotFoundException
     return UserSchemaResponse.from_orm(user).json(), http.HTTPStatus.OK
 
 
@@ -36,6 +38,9 @@ def single_user(scope, id: int):
 def new_user(scope):
     if "create user" not in scope:
         raise AccessDeniedException
+    user = user_repository.find_by_email(request.json['email'])
+    if user:
+        raise UserExistsException
     user = user_repository.create(UserSchemaRequest.parse_obj(request.json))
     cart_repository.create(CartSchema.parse_obj({'user_id': user.id}))
     return UserSchemaResponse.from_orm(user).json(), http.HTTPStatus.CREATED
@@ -46,6 +51,9 @@ def new_user(scope):
 def update_user(scope, id: int):
     if "update user" not in scope:
         raise AccessDeniedException
+    user = user_repository.find(id)
+    if not user:
+        raise UserNotFoundException
     user = user_repository.update(id, UserSchemaRequest.parse_obj(request.json))
     return UserSchemaResponse.from_orm(user).json(), http.HTTPStatus.ACCEPTED
 
@@ -55,5 +63,8 @@ def update_user(scope, id: int):
 def delete_user(scope, id: int):
     if "delete user" not in scope:
         raise AccessDeniedException
+    user = user_repository.find(id)
+    if not user:
+        raise UserNotFoundException
     user_repository.delete(id)
     return jsonify({"message": "User was successfully deleted"}), http.HTTPStatus.NO_CONTENT
